+++ pkgs/objective_c/test/interface_lists_test.dart
@@ -68,9 +68,24 @@ void main() {
           'generated categories', objCBuiltInCategories, allCategoryNames);
     });
 
+    test('All code genned globals are included in the list', () {
+      final allGlobals = findBindings(RegExp(r'^\w+ get (\w+) =>'));
+      expectSetsEqual('generated globals', objCBuiltInGlobals, allGlobals);
+    });
+
     test('No stubs', () {
       final stubRegExp = RegExp(r'\Wstub\W');
       expect(bindings.where(stubRegExp.hasMatch).toList(), <String>[]);
+    });
+
+    test('No automatically renamed classes', () {
+      // All automatically renamed classes or enums should be given an explicit
+      // name. Note that we're not checking for renamed extensions, because ObjC
+      // allows categories with identical names (so we can't unambiguously
+      // rename them), and users don't need to refer to the extension by name
+      // anyway.
+      final renameRegExp = RegExp(r'(class|enum) .*\$');
+      expect(bindings.where(renameRegExp.hasMatch).toList(), <String>[]);
     });
   });
 }