+++ pkgs/objective_c/test/setup.dart
@@ -25,6 +25,7 @@ final objCFiles = [
   'src/ns_number.m',
   'src/objective_c.m',
   'src/objective_c_bindings_generated.m',
+  'src/observer.m',
   'src/protocol.m',
 ].map(_resolve);
 const objCFlags = [
@@ -97,6 +98,7 @@ void main(List<String> arguments) {
   lib.lookup('DOBJC_runOnMainThread'); // objective_c.m
   lib.lookup('Dart_InitializeApiDL'); // dart_api_dl.c
   lib.lookup('OBJC_CLASS_\$_DOBJCDartProtocol'); // protocol.m
+  lib.lookup('OBJC_CLASS_\$_DOBJCObservation'); // observer.m
   // objective_c_bindings_generated.m
   lib.lookup('_ObjectiveCBindings_wrapListenerBlock_ovsamd');