+++ pkgs/swift2objc/lib/src/ast/declarations/compounds/members/property_declaration.dart
@@ -33,26 +33,38 @@ class PropertyDeclaration extends AstNode
   @override
   bool async;
 
+  bool mutating;
+
   bool hasSetter;
 
   PropertyStatements? getter;
   PropertyStatements? setter;
 
+  bool unowned;
+
+  bool weak;
+
+  bool lazy;
+
   bool isStatic;
 
-  PropertyDeclaration({
-    required this.id,
-    required this.name,
-    required this.type,
-    this.hasSetter = false,
-    this.isConstant = false,
-    this.hasObjCAnnotation = false,
-    this.getter,
-    this.setter,
-    this.isStatic = false,
-    this.throws = false,
-    this.async = false,
-  })  : assert(!(isConstant && hasSetter)),
+  PropertyDeclaration(
+      {required this.id,
+      required this.name,
+      required this.type,
+      this.hasSetter = false,
+      this.isConstant = false,
+      this.hasObjCAnnotation = false,
+      this.getter,
+      this.setter,
+      this.isStatic = false,
+      this.throws = false,
+      this.async = false,
+      this.unowned = false,
+      this.weak = false,
+      this.lazy = false,
+      this.mutating = false})
+      : assert(!(isConstant && hasSetter)),
         assert(!(hasSetter && throws));
 
   @override