+++ pkgs/hooks/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.19.5
+
+* Stop leaking unexported symbols.
+
+## 0.19.4
+
+* Add doc comments to all public members.
+
 ## 0.19.3
 
 * Mark this package as in preview.